backend/requirements.txt
requests
anthropic
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
scikit-learn
numpy
fastapi 
uvicorn
openai
motor 
pymongo 
pydantic[email] 
python-jose[cryptography] 
passlib[bcrypt] 
pydantic-settings